requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-adminlte-3==0.1.6
django-autoslug-iplweb==1.9.5
django-cabinet==0.11.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-imagefield==0.13.0
django-js-asset==1.2.2
django-mapbox-location-field==1.7.1
django-mptt==0.11.0
django-parler==2.2
django-ranged-response==0.2.0
django-rosetta==0.9.3
django-simple-captcha==0.5.13
django-tree-queries==0.4.3
idna==2.10
Pillow==8.1.0
polib==1.1.1
python-dotenv==0.16.0
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
validate-email==1.3
